telegram_business_bot/requirements.txt
aiogram>=3.15.0
apscheduler>=3.10.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
